refactor filmavit to neighbor_vit
